fix vlan

See merge request cloudsoft/oneportal-frontend!369

+++ apps/app-smart-cloud/src/app/pages/vlan/delete/delete-port/delete-port.component.ts
@@ -12,11 +12,14 @@ export class DeletePortComponent {
   @Input() region: number
   @Input() project: number
   @Input() id: string
+  @Input() attach: any
   @Output() onOk = new EventEmitter()
   @Output() onCancel = new EventEmitter()
 
   isVisibleDeletePort: boolean = false
   isLoadingDeletePort: boolean = false
+
+  noti: boolean = false;
 
   constructor(@Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService,
               private notification: NzNotificationService,
@@ -36,16 +39,22 @@ export class DeletePortComponent {
 
   handleOkDeletePort() {
     this.isLoadingDeletePort = true
-    this.vlanService.deletePort(this.id, this.region, this.project).subscribe(data => {
-      console.log('delete', data)
-      this.isVisibleDeletePort = false
-      this.isLoadingDeletePort = false
-      this.notification.success('Thành công', 'Yêu cầu xoá Port thành công')
-      setTimeout(() => {this.onOk.emit(data)}, 1500)
-    }, error => {
-      this.isVisibleDeletePort = false
-      this.isLoadingDeletePort = false
-      this.notification.error('Thất bại', 'Yêu cầu xoá Port thất bại')
-    })
+    if(this.attach == undefined || this.attach == null) {
+      this.noti = false
+      this.vlanService.deletePort(this.id, this.region, this.project).subscribe(data => {
+        console.log('delete', data)
+        this.isVisibleDeletePort = false
+        this.isLoadingDeletePort = false
+        this.notification.success('Thành công', 'Yêu cầu xoá Port thành công')
+        setTimeout(() => {this.onOk.emit(data)}, 1500)
+      }, error => {
+        this.isVisibleDeletePort = false
+        this.isLoadingDeletePort = false
+        this.notification.error('Thất bại', 'Yêu cầu xoá Port thất bại')
+      })
+    } else {
+      this.noti = true
+    }
+
   }
 }
